+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to Sourcegraph Server and Data Center are documented in this
 
 * `gitMaxConcurrentClones` now also limits the concurrency of updates to repos in addition to the initial clone.
 * In the GraphQL API, `site.users` has been renamed to `users`, `site.orgs` has been renamed to `organizations`, and `site.repositories` has been renamed to `repositories`.
+* An authentication provider must be set in site configuration (see [authentication provider documentation](https://about.sourcegraph.com/docs/config/authentication). Previously the server defaulted to builtin auth if none was set.
 
 ### Fixed